+++ comb_spec_searcher/__init__.py
@@ -16,7 +16,7 @@ from .strategies import (
     VerificationStrategy,
 )
 
-__version__ = "4.1.0"
+__version__ = "4.2.0"
 
 __all__ = [
     "CombinatorialSpecificationSearcher",