+++ comb_spec_searcher/strategies/constructor/cartesian.py
@@ -349,6 +349,7 @@ class CartesianProduct(Constructor[CombinatorialClassType, CombinatorialObjectTy
                     subsampler(n=extra_params.pop("n"), **extra_params)
                     for subsampler, extra_params in zip(subsamplers, extra_parameters)
                 )
+        raise RuntimeError("Function did not return")
 
     @staticmethod
     def get_eq_symbol() -> str:
@@ -545,6 +546,7 @@ class Quotient(Constructor[CombinatorialClassType, CombinatorialObjectType]):
 
     def _terms_to_poly(self, terms: Terms) -> Union[sympy.Poly, int]:
         variables = tuple(sympy.var(f"k_{i}") for i in range(self._num_parent_params))
+        res: Union[sympy.Poly, int]
         if variables:
             res = sympy.Poly(sympy.sympify("0"), *variables)
         else:
@@ -590,6 +592,8 @@ class Quotient(Constructor[CombinatorialClassType, CombinatorialObjectType]):
     def get_terms(
         self, parent_terms: Callable[[int], Terms], subterms: SubTerms, n: int
     ) -> Terms:
+        if n < self._min_sizes[self.idx]:
+            return Counter()
         new_terms: Terms = Counter()
         children_subterms = (
             subterms[1 : self.idx + 1] + (parent_terms,) + subterms[self.idx + 1 :]