+++ CHANGELOG.md
@@ -5,6 +5,32 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `VerificationRule.from_dict` reapplies strategies. 
+
+### Fixed
+- Bug in `CombinatorialSpecification.expand_comb_class`, 
+  it will now unpack all of the `EquivalencePathRule`s. 
+- Bug in `Quotient.get_terms`. Look at reliance profile 
+  to check for initial conditions, so as to avoid max 
+  recursion error.
+- `CombinatorialSpecification.get_genf` will not try to 
+  solve systems of equations with catalytic variables.
+- Bug in `DisjointUnion.build_param_map`. If two params 
+  map to the same, then these shouldn't sum. 
+- `EquivalencePathRule` can not accept `Complement` rules 
+  where two parameters map to the same parameter as this will 
+  result in more than permuting labels.
+
+## [4.2.0] 2023-01-18
+### Changed
+- Update dependency on psutil from 5.8 to 5.9.4
+- Update dependency on sympy from 1.9 to 1.10.1
+- Update dependency on pympler from 0.9 to 1.0.1
+- Update dependency on requests from 2.26.0 to 2.28.1
+- Update dependency on typing-extensions from 4.0.0 to 4.4.0
+- Update dependency on tabulate from 0.8.9 to 0.9.0
+- Updated dependencies in the tox file
 
 ## [4.1.0] 2022-01-17
 ### Added